refacto: Administration (fixes #70)
Change net.http to fasthttp
