+++ orchestrator/starter.go
@@ -1,7 +1,7 @@
 package orchestrator
 
 import (
-	"github.com/forsam-education/cerberus/admin"
+	"github.com/forsam-education/cerberus/administration"
 	"github.com/forsam-education/cerberus/database"
 	"github.com/forsam-education/cerberus/proxy"
 	"github.com/forsam-education/cerberus/state"
@@ -59,8 +59,7 @@ func StartOrchestrator() {
 	waitgroup.Add(1)
 
 	go proxy.StartServer(&waitgroup)
-	//go administration.StartServer(context.Background(), &waitgroup)
-	go admin.StartServer(&waitgroup)
+	go administration.StartServer(&waitgroup)
 
 	waitgroup.Wait()
